Added Story upto Day43 and some Tutorial Edits (#1)
* Added Story upto Day43 and some Tutorial Edits

There are a couple typos in story and it hasn't been proofread through a second time yet. Specific typos would be on one day there is a '/' instead of a '.' at the end of a sentence and the edit to change some text to italics looks odd in game because of the speed the text is displayed at. (Italics problem should be on a line relating to grains of sand on the floor shining with light)

* Updated Translation Progress

* Updated translation_progress with Story tracking.